requirements.txt
arabic-reshaper
python-bidi
gradio 
networkx 
matplotlib 
pandas
stanza
Tashaphyne
rich
pyvis